requirements.txt
streamlit==1.28.0
pillow==10.3.0
qrcode==7.4.2